bin: add kaui command line tool
The kaui command line tool can now be used to setup
a brand new Rails application with Kaui automatically.
It can probably also be used to setup Kaui as a mounted
engine in an existing application, although this hasn't been
tested yet.

The install process is trivial:
* it adds kaui as a dependency in the Rails Gemfile
* it then runs: rails generate kaui:install

kaui:install is a new Rails generator which will copy template files
for all assets and add Kaui specific routes.

Documentation on how to use it can be found in the Getting Started guide:
http://killbilling.org/start.html

The skeleton code for both the command line tool as well as the Rails generator
was inspired from http://spreecommerce.com. Thanks!

Signed-off-by: Pierre-Alexandre Meyer <pierre@ning.com>