If an account has multiple bundles with the same externalKey, only one can be active.
Currently the code (get_bundle_by_key) returns that latest one from the list, which might not match the active bundle.
The fix is to check if there is an active one, and if not return the one for which latest non ADD_ON subscription was created.

This should fix problem is KAUI where we reort wrong information.