Re-implement logic around browser back button to end up on the right admin_tenant config TAB. See #72
The new logic removes the extra back button and relies on user clicking browser back button instead. The code works
as expected, but because we rely on setting the new url from the view (with the right query parameter), this forces
each time we swicth page to relaod the page (with all the calls to KB). this is quite costly.

Note: I also tried a variation where we don't set the query param from the view but instead modify the referrer but this does not
work because browser button will not look into that (and instead use browser history). One option would be to modify the browser
history to insert the correct entry but that seems complicated and i am affraid different browser will behave differently)