payments: display last transaction status in payments table
Unfortunately, UNKNOWN and PLUGIN_FAILURE transactions share the same payment
state name. When searching for UNKNOWN payments only, we used to filter out
PLUGIN_FAILURE transactions from the search results. This was quite confusing
as the table statistics (number of records, etc.) didn't match what was on screen.

We now keep all results, but display the last transaction status in the table for
convenience.

Signed-off-by: Pierre-Alexandre Meyer <pierre@mouraf.org>